+++ tests/test_mlx5_raw_wqe.py
@@ -1,46 +1,20 @@
 # SPDX-License-Identifier: (GPL-2.0 OR Linux-OpenIB)
 # Copyright (c) 2021 Nvidia, Inc. All rights reserved. See COPYING file
 
-import unittest
-import random
-import errno
-
-
-from pyverbs.providers.mlx5.mlx5dv import Mlx5Context, Mlx5DVContextAttr, Mlx5DVQPInitAttr, \
-    Mlx5QP, Mlx5DVCQInitAttr, Mlx5CQ, Wqe, WqeDataSeg, WqeCtrlSeg
-from pyverbs.pyverbs_error import PyverbsRDMAError, PyverbsUserError, PyverbsError
+from pyverbs.providers.mlx5.mlx5dv import Wqe, WqeDataSeg, WqeCtrlSeg
+from pyverbs.pyverbs_error import PyverbsError
 import pyverbs.providers.mlx5.mlx5_enums as dve
-from tests.mlx5_base import Mlx5RDMATestCase
-from pyverbs.qp import QPInitAttrEx, QPCap
-from pyverbs.cq import CqInitAttrEx
-from tests.base import RCResources
+from tests.mlx5_base import Mlx5RDMATestCase, Mlx5RcResources
+from pyverbs.qp import QPCap
 from pyverbs.wr import SGE
-from pyverbs.mr import MR
 import pyverbs.enums as e
 import tests.utils as u
 
 
-class Mlx5RawWqeResources(RCResources):
-    def __init__(self, dev_name, ib_port, gid_index):
+class Mlx5RawWqeResources(Mlx5RcResources):
+    def create_send_ops_flags(self):
         self.dv_send_ops_flags = dve.MLX5DV_QP_EX_WITH_RAW_WQE
         self.send_ops_flags = e.IBV_QP_EX_WITH_SEND
-        super().__init__(dev_name, ib_port, gid_index)
-
-    def create_context(self):
-        mlx5dv_attr = Mlx5DVContextAttr()
-        try:
-            self.ctx = Mlx5Context(mlx5dv_attr, name=self.dev_name)
-        except PyverbsUserError as ex:
-            raise unittest.SkipTest(f'Could not open mlx5 context ({ex})')
-        except PyverbsRDMAError:
-            raise unittest.SkipTest('Opening mlx5 context is not supported')
-
-    def create_qp_init_attr(self):
-        comp_mask = e.IBV_QP_INIT_ATTR_PD | e.IBV_QP_INIT_ATTR_SEND_OPS_FLAGS
-        return QPInitAttrEx(cap=self.create_qp_cap(), pd=self.pd, scq=self.cq,
-                            rcq=self.cq, qp_type=e.IBV_QPT_RC,
-                            send_ops_flags=self.send_ops_flags,
-                            comp_mask=comp_mask)
 
     def create_qp_cap(self):
         """
@@ -50,35 +24,6 @@ class Mlx5RawWqeResources(RCResources):
         :return:
         """
         return QPCap(max_send_wr=4, max_recv_wr=4, max_recv_sge=2, max_send_sge=2)
-
-
-    def create_qps(self):
-        try:
-            qp_init_attr = self.create_qp_init_attr()
-            comp_mask = dve.MLX5DV_QP_INIT_ATTR_MASK_QP_CREATE_FLAGS | \
-                    dve.MLX5DV_QP_INIT_ATTR_MASK_SEND_OPS_FLAGS
-            attr = Mlx5DVQPInitAttr(comp_mask=comp_mask, send_ops_flags=self.dv_send_ops_flags)
-            qp = Mlx5QP(self.ctx, qp_init_attr, attr)
-            self.qps.append(qp)
-            self.qps_num.append(qp.qp_num)
-            self.psns.append(random.getrandbits(24))
-        except PyverbsRDMAError as ex:
-            if ex.error_code == errno.EOPNOTSUPP:
-                raise unittest.SkipTest('Create Mlx5DV QP is not supported')
-            raise ex
-
-    def create_cq(self):
-        """
-        Initializes self.cq with a dv_cq
-        :return: None
-        """
-        dvcq_init_attr = Mlx5DVCQInitAttr()
-        try:
-            self.cq = Mlx5CQ(self.ctx, CqInitAttrEx(), dvcq_init_attr)
-        except PyverbsRDMAError as ex:
-            if ex.error_code == errno.EOPNOTSUPP:
-                raise unittest.SkipTest('Create Mlx5DV CQ is not supported')
-            raise ex
 
 
 class RawWqeTest(Mlx5RDMATestCase):
